+++ src/components/sidebar/__spec__.js
@@ -1,38 +1,33 @@
 import React from 'react';
-import TestUtils from 'react-dom/test-utils';
-import { shallow, mount } from 'enzyme';
+import { shallow, mount, ReactWrapper } from 'enzyme';
 import Browser from '../../utils/helpers/browser';
 import { Sidebar } from './sidebar';
 import Textbox from './../textbox';
+import Portal from './../portal';
 import Icon from './../icon';
-import { elementsTagTest, rootTagTest } from '../../utils/helpers/tags/tags-specs';
 
 describe('Sidebar', () => {
-  let instance, leftInstance, spy;
+  let wrapper, portalContent, spy;
 
   beforeEach(() => {
     spy = jasmine.createSpy();
 
-    instance = TestUtils.renderIntoDocument(
+    wrapper = shallow(
       <Sidebar
-        onCancel={ spy }
-        open={ true }
+        open
+        title='Test'
         className='custom-class'
+        data-role='baz'
+        data-element='bar'
+        onCancel={ spy }
       >
         <Textbox />
         <Textbox />
         <Textbox />
       </Sidebar>
     );
-
-    leftInstance = TestUtils.renderIntoDocument(
-      <Sidebar
-        onCancel={ spy }
-        open={ true }
-        size='small'
-        position='left'
-        enableBackgroundUI={ true }
-      />
+    portalContent = new ReactWrapper(
+      wrapper.find(Portal).prop('children')
     );
   });
 
@@ -53,7 +48,7 @@ describe('Sidebar', () => {
     describe('when the Sidebar is open', () => {
       beforeEach(() => {
         wrapper = mount(
-          <Sidebar open={ true } onCancel={ spy } />
+          <Sidebar open onCancel={ spy } />
         );
         instance = wrapper.instance();
       });
@@ -97,81 +92,53 @@ describe('Sidebar', () => {
   });
 
   describe('sidebarClasses', () => {
-    it('returns a base sidebar class', () => {
-      expect(instance.sidebarClasses).toMatch('carbon-sidebar__sidebar carbon-sidebar__sidebar--right carbon-sidebar__sidebar--medium');
-    });
-
-    it('returns a position modifier class based on props', () => {
-      expect(leftInstance.sidebarClasses).toMatch('carbon-sidebar__sidebar--left');
-    });
-
-    it('returns a size modifier class based on props', () => {
-      expect(leftInstance.sidebarClasses).toMatch('carbon-sidebar__sidebar--small');
-    });
-  });
-
-  describe('mainClasses', () => {
-    it('sets the sidebar class', () => {
-      expect(leftInstance.mainClasses).toEqual('carbon-sidebar');
-    });
-
-    it('appends any additional passed classes', () => {
-      expect(instance.mainClasses).toEqual('carbon-sidebar custom-class');
-    });
-  });
-
-  describe('sidebarHTML', () => {
-    it('returns the rendered sidebar', () => {
-      expect(TestUtils.findRenderedDOMComponentWithClass(instance, 'carbon-sidebar__sidebar')).toBeTruthy();
-    });
-
-    it('renders a close icon', () => {
-      let icon = TestUtils.findRenderedComponentWithType(instance, Icon);
-      expect(icon.props.type).toEqual('close');
-    });
-
-    it('renders children within the sidebar', () => {
-      let textboxes = TestUtils.scryRenderedDOMComponentsWithClass(instance, 'carbon-textbox');
-      expect(textboxes.length).toEqual(3);
+    it('returns a base sidebar', () => {
+      expect(wrapper).toMatchSnapshot();
     });
   });
 
   describe('render', () => {
-    describe('when sidebar is open', () => {
-      it('renders a parent div', () => {
-        expect(TestUtils.findRenderedDOMComponentWithClass(instance, 'carbon-sidebar')).toBeTruthy();
-      });
-
-      it('renders a background div', () => {
-        expect(TestUtils.findRenderedDOMComponentWithClass(instance, 'carbon-modal__background')).toBeTruthy();
-      });
-
-      it('renders a sidebar', () => {
-        expect(TestUtils.findRenderedDOMComponentWithClass(instance, 'carbon-sidebar__sidebar')).toBeTruthy();
+    describe('when sidebar is closed', () => {
+      it('sets all the correct classes', () => {
+        wrapper = shallow(<Sidebar onCancel={ spy } />);
+        portalContent = new ReactWrapper(
+          wrapper.find(Portal).prop('children')
+        );
+        expect(portalContent.find('.carbon-sidebar').text()).toEqual('');
       });
     });
 
-    describe('when sidebar is closed', () => {
-      beforeEach(() => {
-        instance = TestUtils.renderIntoDocument(
+    describe('when enableBackgroundUI is enabled', () => {
+      it('sets all the correct classes', () => {
+        wrapper = shallow(
           <Sidebar
+            open
+            enableBackgroundUI
+            size='small'
+            position='left'
             onCancel={ spy }
-            open={ false }
-          >
-          </Sidebar>
+          />
         );
+        portalContent = new ReactWrapper(
+          wrapper.find(Portal).prop('children')
+        );
+        expect(portalContent.find('.carbon-modal__background').length).toEqual(0);
+        expect(portalContent.find('.carbon-sidebar__sidebar--left').length).toEqual(1);
+        expect(portalContent.find('.carbon-sidebar__sidebar--small').length).toEqual(1);
       });
+    });
 
-      it('renders a parent div', () => {
-        expect(TestUtils.findRenderedDOMComponentWithClass(instance, 'carbon-sidebar')).toBeTruthy();
-      });
-
-      it('does not render a background div', () => {
-        expect(TestUtils.scryRenderedDOMComponentsWithClass(instance, 'carbon-sidebar__background').length).toEqual(0);
-      });
-
-      it('does not render a sidebar', () => {
-        expect(TestUtils.scryRenderedDOMComponentsWithClass(instance, 'carbon-sidebar__sidebar').length).toEqual(0);
+    describe('when there is no onCancel prop', () => {
+      it('should not have a close button', () => {
+        wrapper = shallow(
+          <Sidebar
+            open
+          />
+        );
+        portalContent = new ReactWrapper(
+          wrapper.find(Portal).prop('children')
+        );
+        expect(portalContent.find('.carbon-sidebar__close').length).toEqual(0);
       });
     });
   });
@@ -179,57 +146,10 @@ describe('Sidebar', () => {
   describe('Behaviour', () => {
     describe('clicking the close icon sidebar', () => {
       it('closes the sidebar', () => {
-        let icon = TestUtils.findRenderedDOMComponentWithClass(instance, 'carbon-sidebar__close-icon');
-        TestUtils.Simulate.click(icon);
+        let icon = portalContent.find('.carbon-sidebar__close-icon');
+        icon.simulate('click');
         expect(spy).toHaveBeenCalled();
       });
-    });
-
-    describe('when there is no onCancel prop', () => {
-      beforeEach(() => {
-        instance = TestUtils.renderIntoDocument(
-          <Sidebar open={ true }>
-            <Textbox />
-          </Sidebar>
-        );
-      });
-
-      it('does not render a close icon', () => {
-        let icon = TestUtils.scryRenderedDOMComponentsWithClass(instance, 'carbon-sidebar__close-icon');
-        expect(icon.length).toEqual(0);
-      });
-    });
-  });
-
-  describe("tags", () => {
-    describe("on component", () => {
-      let wrapper = shallow(
-        <Sidebar
-          data-element='bar'
-          onCancel={ () => {} }
-          onConfirm={ () => {} }
-          data-role='baz'
-        />
-      );
-
-      it('include correct component, element and role data tags', () => {
-        rootTagTest(wrapper, 'sidebar', 'bar', 'baz');
-      });
-    });
-
-    describe("on internal elements", () => {
-      let wrapper = shallow(
-        <Sidebar
-          open={ true }
-          title='Test'
-          onCancel={ () => {} }
-          onConfirm={ () => {} }
-        />
-      );
-
-      elementsTagTest(wrapper, [
-        'close'
-      ]);
     });
   });
 });